Rename JSON to FList
